+++ package.json
@@ -23,6 +23,7 @@
     "date-fns": "^2.30.0",
     "flowbite": "^2.1.1",
     "flowbite-react": "^0.6.4",
+    "lodash.debounce": "^4.0.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.48.2",
@@ -35,6 +36,7 @@
     "yup": "^1.3.2"
   },
   "devDependencies": {
+    "@types/lodash.debounce": "^4.0.9",
     "@types/node": "^20.8.2",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
