+++ src/components/NewsFIlterManager/assistants/constants.ts
@@ -1,0 +1,53 @@
+export const materialTypes = [
+  'Addendum',
+  'An Analysis',
+  'An Appraisal',
+  'Archives',
+  'Article',
+  'Banner',
+  'Biography',
+  'Birth Notice',
+  'Blog',
+  'Briefing',
+  'Caption',
+  'Chronology',
+  'Column',
+  'Correction',
+  'Economic Analysis',
+  'Editorial',
+  'Editorial Cartoon',
+  "Editors' Note",
+  'First Chapter',
+  'Front Page',
+  'Glossary',
+  'Interactive Feature',
+  'Interactive Graphic',
+  'Interview',
+  'Letter',
+  'List',
+  'Marriage Announcement',
+  'Military Analysis',
+  'News',
+  'News Analysis',
+  'Newsletter',
+  'Obituary',
+  'Obituary (Obit)',
+  'Op-Ed',
+  'Paid Death Notice',
+  'Postscript',
+  'Premium',
+  'Question',
+  'Quote',
+  'Recipe',
+  'Review',
+  'Schedule',
+  'SectionFront',
+  'Series',
+  'Slideshow',
+  'Special Report',
+  'Statistics',
+  'Summary',
+  'Text',
+  'Video',
+  'Web Log',
+];
