+++ src/components/NewsFIlterManager/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useFilterNews } from './useFilterNews';
