+++ src/components/NewsFIlterManager/subcomponents/Calendar/subcomponents/CalendarContent.tsx
