+++ src/components/NewsFIlterManager/subcomponents/Calendar/subcomponents/index.ts
