+++ src/components/NewsFIlterManager/subcomponents/SearchBlock.tsx
