+++ src/components/NewsFIlterManager/subcomponents/index.ts
@@ -1,0 +1,2 @@
+export { default as FiltersBlock } from './FiltersBlock';
+export { default as SearchBlock } from './SearchBlock';
