+++ src/components/PageScrollController/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useScrollController } from './useScrollController';
