+++ src/components/index.ts
@@ -1,19 +1,15 @@
 export { default as AccountMenu } from './AccountMenu';
-export { default as Accordeon } from './Accordeon';
 export { default as Auth } from './Auth';
 export { default as AuthModal } from './AuthModal/AuthModal';
-export { default as Calendar } from './Calendar/Calendar';
-export { default as FiltersBlock } from './FiltersBlock';
 export { default as Hero } from './Hero';
 export { default as Loader } from './Loader/Loader';
 export { default as Menu } from './Menu';
-export { default as NewsFilterManager } from './NewsFilterManager';
+export { default as NewsFilterManager } from './NewsFIlterManager/NewsFilterManager';
 export { default as NewsItem } from './NewsItem';
 export { default as NewsList } from './NewsList';
-export { default as PageScrollController } from './PageScrollController';
+export { default as PageScrollController } from './PageScrollController/PageScrollController';
 export { default as Pagination } from './Pagination';
 export { default as PlugImage } from './PlugImage/PlugImage';
-export { default as SearchBlock } from './SearchBlock';
 export { default as ThemeSwitcher } from './ThemeSwitcher';
 export { default as VoteButton } from './VoteButton';
 export { default as WeatherBlock } from './WeatherBlock/WeatherBlock';
