+++ src/helpers/index.ts
@@ -1,5 +1,3 @@
-export { default as applyCrossFilters } from './applyCrossFilters';
-export { default as organiseNewsByMonth } from './organiseNewsByMonth';
 export { default as rebuildNewsArray } from './rebuildNewsArray';
 export { default as updateLocaleStorage } from './updateLocaleStorage';
 export * from './dateTimeHelpers';
