+++ src/hooks/index.ts
@@ -1,12 +1,10 @@
-export { default as useWindowWidth } from './useWindowWidth';
-export { default as usePopUp } from './usePopUp';
-export { default as useHeaderStyles } from './useHeaderStyles';
-export { default as usePagination } from './usePagination';
 export { default as useActiveLinks } from './useActiveLinks';
 export { default as useAdditionalRequest } from './useAdditionalRequest';
-export { default as useFilterNews } from './useFilterNews';
+export { default as useAuthCollector } from './useAuthCollector';
 export { default as useChooseRenderingNews } from './useChooseRenderingNews';
+export { default as useFilterCollector } from './useFilterCollector';
+export { default as useHeaderStyles } from './useHeaderStyles';
 export { default as useNewsAPICollector } from './useNewsAPICollector';
 export { default as useNewsDBCollector } from './useNewsDBCollector';
-export { default as useAuthCollector } from './useAuthCollector';
-export { default as useFilterCollector } from './useFilterCollector';
+export { default as usePopUp } from './usePopUp';
+export { default as useWindowWidth } from './useWindowWidth';
