+++ src/pages/ArchivePage/assistants/index.ts
@@ -1,0 +1,1 @@
+export { default as organiseNewsByMonth } from './utils';
