+++ src/pages/HomePage/hooks/usePagination.ts
@@ -1,10 +1,7 @@
 import { useEffect, useState } from 'react';
 
 import { PartialVotedNewsArray } from 'types';
-
-import useWindowWidth from './useWindowWidth';
-import useNewsAPICollector from './useNewsAPICollector';
-import useFilterCollector from './useFilterCollector';
+import { useWindowWidth, useNewsAPICollector, useFilterCollector } from 'hooks';
 
 const usePagination = (rebuildedNews: PartialVotedNewsArray) => {
   const { breakpointsForMarkup } = useWindowWidth() || {
