+++ src/pages/ReadPage.tsx
@@ -2,7 +2,8 @@ import React, { FC, useEffect, useLayoutEffect, useState } from 'react';
 
 import { useNewsDBCollector } from 'hooks';
 
-import { Accordeon, Loader, NewsList, PlugImage } from 'components';
+import { Loader, NewsList, PlugImage } from 'components';
+import { Accordeon } from 'ui';
 
 const ReadPage: FC<{}> = () => {
   const [changesHappened, setChangesHappened] = useState<boolean>(false);
