+++ src/pages/index.ts
@@ -1,7 +1,7 @@
 export { default as AccountPage } from './AccountPage';
 export { default as AccountManagePage } from './AccountManagePage/AccountManagePage';
-export { default as ArchivePage } from './ArchivePage';
+export { default as ArchivePage } from './ArchivePage/ArchivePage';
 export { default as ErrorPage } from './ErrorPage';
 export { default as FavouritePage } from './FavouritePage';
-export { default as HomePage } from './HomePage';
+export { default as HomePage } from './HomePage/HomePage';
 export { default as ReadPage } from './ReadPage';
