+++ src/ui/index.ts
@@ -1,3 +1,4 @@
+export { default as Accordeon } from './Accordeon';
 export { default as CommonMenu } from './CommonMenu';
 export { default as Dropdown } from './Dropdown';
 export { default as Modal } from './Modal';
